+++ back-end/app/Http/Resources/TransporterResource.php
@@ -17,7 +17,7 @@ class TransporterResource extends JsonResource
         return [
             'id'=>$this->id,
             'title'=>$this->title,
-            'price'=>(float)$this->price,
+            'price'=>(int)$this->price,
             'image'=>$this->image != null ? asset('assets/images/'.$this->image) : null,
         ];
     }
